requirements.txt
annotated-types==0.6.0
anthropic==0.21.3
anyio==4.3.0
blis==0.7.11
boto3==1.34.76
botocore==1.34.76
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
Deprecated==1.2.14
detoxify==0.5.2
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.3
fsspec==2024.3.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
importlib-metadata==7.0.0
Jinja2==3.1.3
jmespath==1.0.1
langcodes==3.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai==1.16.1
opentelemetry-api==1.24.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
packaging==24.0
phonenumbers==8.13.33
preshed==3.0.9
presidio-analyzer==2.2.354
protobuf==4.25.3
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
s3transfer==0.10.1
safetensors==0.4.2
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.3
tldextract==5.1.2
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
triton==2.2.0
typer==0.9.4
typing_extensions==4.10.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
wrapt==1.16.0
zipp==3.18.1
